add authorization to frontend

+++ frontend/src/components/MessageDetails.js
@@ -1,14 +1,21 @@
 import { useMessagesContext } from "../hooks/useMessagesContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const MessageDetails = ({ message }) => {
   const { dispatch } = useMessagesContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("/api/messages/" + message._id, {
       method: "DELETE",
+      headers: { Authorization: `Bearer ${user.token}` },
     });
     const json = await response.json();
